day6/project/aws-sql-student-system/requirements.txt
Flask-Mail
psycopg2-binary
python-dotenv
